requirements.txt
gradio>=4.0.0
python-dotenv>=1.0.0
torch>=2.0.0
numpy>=1.24.0
transformers>=4.36.0
torchaudio>=2.0.0
boto3>=1.34.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
replicate>=0.10.0
Pillow>=10.0.0
requests>=2.28.0